feat: move components from locales to utils
