--- src/components/BannerSlider/Banner.tsx
@@ -1,7 +1,0 @@
-import styles from './Banner.module.sass'
-
-export default function Home() {
-  return (
-    <h1 className={styles.title}></h1>
-  )
-}
