+++ src/components/Header/DropDown/ProfileDropDown/ProfileDropDown.tsx
@@ -2,18 +2,18 @@ import {FC} from "react";
 import styles from './ProfileDropDown.module.sass';
 import CustomButton from "/src/UI/CustomButton/CustomButton";
 import {profileDropDownCards} from "/src/locales/profileDropDownData";
-import {profileLinks} from "/src/components/Header/DropDown/ProfileDropDown/ProfileDropDown.utils";
+import {cardsIcons, profileLinks} from "/src/components/Header/DropDown/ProfileDropDown/ProfileDropDown.utils";
 
 const ProfileDropDown: FC = () => {
     return (
         <div className={styles.profile__dropdown_container}>
             <div className={styles.profile__main}>
                 {
-                    profileDropDownCards.icons.map((icon, i) => {
+                    profileDropDownCards.cardsText.map((text, i) => {
                             return (
                                 <a className={styles.profile__main_card} href={profileLinks[i]} target='_blank' key={i}>
-                                    <div className={styles.card__icon}>{icon}</div>
-                                    <div className={styles.card__text}>{profileDropDownCards.cardsText[i]}</div>
+                                    <div className={styles.card__icon}>{cardsIcons[i]}</div>
+                                    <div className={styles.card__text}>{text}</div>
                                 </a>
                             )
                         }
