--- src/components/Header/DropDown/ProfileDropDown/ProfileDropDown.utils.ts
@@ -1,9 +1,0 @@
-export const profileLinks = [
-    "https://www.ivi.ru/profile/purchases",
-    "https://www.ivi.ru/profile/favorites",
-    "https://www.ivi.ru/profile/watched",
-    "https://www.ivi.ru/profile/subscriptions",
-    "https://www.ivi.ru/profile/subscriptions",
-    "https://www.ivi.ru/profile/cards",
-    "https://www.ivi.ru/profile/cards",
-]
