+++ src/locales/profileDropDownData.ts
@@ -1,0 +1,3 @@
+export const profileDropDownCards = {
+    cardsText: ['Покупки', 'Смотреть позже', 'История просмотров', 'Подписки', 'Активация сертификата', 'Вход по коду', 'Способы оплаты'],
+}
